Seperate getCompletionNode from getNodeAtPosition
