+++ packages/compiler/server/serverlib.ts
@@ -50,6 +50,7 @@ import {
 } from "vscode-languageserver/node.js";
 import { defaultConfig, findCadlConfigPath, loadCadlConfigFile } from "../config/config-loader.js";
 import { CadlConfig } from "../config/types.js";
+import { codePointBefore, isIdentifierContinue } from "../core/charcode.js";
 import {
   compilerAssert,
   createSourceFile,
@@ -674,9 +675,7 @@ export function createServer(host: ServerHost): Server {
 
   async function getSignatureHelp(params: SignatureHelpParams): Promise<SignatureHelp | undefined> {
     return await compile(params.textDocument, (program, document, file) => {
-      const nodeAtPosition = getNodeAtPosition(file, document.offsetAt(params.position), {
-        resolveExact: true,
-      });
+      const nodeAtPosition = getNodeAtPosition(file, document.offsetAt(params.position));
       const data = nodeAtPosition && findDecoratorOrParameter(nodeAtPosition);
       if (data === undefined) {
         return undefined;
@@ -769,7 +768,7 @@ export function createServer(host: ServerHost): Server {
       items: [],
     };
     await compile(params.textDocument, async (program, document, file) => {
-      const node = getNodeAtPosition(file, document.offsetAt(params.position));
+      const node = getCompletionNodeAtPosition(file, document.offsetAt(params.position));
       if (node === undefined) {
         addKeywordCompletion("root", completions);
       } else {
@@ -1517,4 +1516,32 @@ function findDecoratorOrParameter(
     current = current.parent;
   }
   return undefined;
+}
+
+/**
+ * Resolve the node that should be auto completed at the given position.
+ * It will try to guess what node it could be as during auto complete the ast might not be complete.
+ * @internal
+ */
+export function getCompletionNodeAtPosition(
+  script: CadlScriptNode,
+  position: number,
+  filter: (node: Node) => boolean = (node: Node) => true
+): Node | undefined {
+  const realNode = getNodeAtPosition(script, position, filter);
+  if (realNode?.kind === SyntaxKind.StringLiteral) {
+    return realNode;
+  }
+  // If we're not immediately after an identifier character, then advance
+  // the position past any trivia. This is done because a zero-width
+  // inserted missing identifier that the user is now trying to complete
+  // starts after the trivia following the cursor.
+  const cp = codePointBefore(script.file.text, position);
+  if (!cp || !isIdentifierContinue(cp)) {
+    const newPosition = skipTrivia(script.file.text, position);
+    if (newPosition !== position) {
+      return getNodeAtPosition(script, newPosition, filter);
+    }
+  }
+  return realNode;
 }
